requirements.txt
allure-pytest==2.13.5
curlify==2.2.1
Faker==30.8.2
pytest==8.3.3
requests==2.32.3
structlog==24.4.0
retrying==1.3.4
pydantic==2.10.1
PyHamcrest==2.1.0
vyper-config==1.1.1
assertpy==1.1